+++ src/config/recommended.ts
@@ -1,10 +1,12 @@
 module.exports = {
 	rules: {
+		'attr-case': 'error',
 		'attr-quotes': 'error',
 		'button-type': 'error',
 		'close-attr': 'error',
 		'close-order': 'error',
 		'deprecated': 'error',
+		'element-case': 'error',
 		'element-name': 'error',
 		'element-permitted-content': 'error',
 		'element-permitted-occurrences': 'error',