+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Features
 
-. add `element-case` rule for validating case of element names.
-. add `attr-case` rule for validating case of attributes.
+- adding `doctype` event, emitted when a doctype is encountered.
+- add `element-case` rule for validating case of element names.
+- add `attr-case` rule for validating case of attributes.
 
 ## 0.9.1 (2018-07-01)